requirements.txt
transformers>=4.6.0,<5.0.0
tokenizers
tqdm
#torch>=1.6.0
torchvision
numpy
scikit-learn
scipy
nltk
sentencepiece
huggingface-hub
pyspellchecker
clearml
pyarrow
spacy==3.2.0
plotly
